Implementing basic_string_view
Adding string_view as a specialization of basic_string_view<char>.

Had to define and declare everything in the header file (that's why the
cc file is gone).

Something I'm not sure about is why operator== (and others) have to be
defined as friend members of the template class in order to be found by
ADL in googletest.